+++ eslint.config.mjs
@@ -1,9 +1,9 @@
 import js from '@eslint/js';
-import path from 'node:path';
-import tsParser from '@typescript-eslint/parser';
 import stylistic from '@stylistic/eslint-plugin';
-import { fileURLToPath } from 'node:url';
+import tsParser from '@typescript-eslint/parser';
+import path from 'node:path';
 import { FlatCompat } from '@eslint/eslintrc';
+import { fileURLToPath } from 'node:url';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -57,6 +57,11 @@ export default [{
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/no-non-null-assertion': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/no-unused-vars': ['error', {
+      argsIgnorePattern: '^_',
+      varsIgnorePattern: '^_',
+      caughtErrorsIgnorePattern: '^_',
+    }],
     '@stylistic/semi': ['warn'],
     '@stylistic/member-delimiter-style': ['warn', {
       multiline: {