+++ src/devices/deviceTypes.ts
@@ -105,22 +105,25 @@ export interface ConfigDevice {
 
 export const Plugs = [
   'EP10', 'EP25', 'HS100', 'HS103', 'HS105', 'HS110', 'KP100', 'KP105', 'KP115', 'KP125', 'KP125M',
-  'KP401', 'P100', 'P110', 'P110M', 'P115', 'P125M', 'P135', 'TP15',
+  'KP401', 'P100', 'P105', 'P110', 'P110M', 'P115', 'P125M', 'P135', 'TP10', 'TP15',
 ];
 export const PowerStrips = [
-  'EP40', 'EP40M', 'HS107', 'HS300', 'KP200', 'KP303', 'KP400', 'P210M', 'P300', 'P304M', 'P306', 'P400M', 'TP25',
+  'EP40', 'EP40M', 'HS107', 'HS300', 'KP200', 'KP303', 'KP400', 'P210M', 'P300', 'P304M', 'P306',
+  'P316M', 'P400M', 'TP25',
 ];
 export const Switches = [
   'ES20M', 'HS200', 'HS210', 'HS220', 'KP405', 'KS200', 'KS200M', 'KS205', 'KS220', 'KS220M', 'KS225',
-  'KS230', 'KS240', 'S500', 'S500D', 'S505', 'S505D', 'TS15',
+  'KS230', 'KS240', 'S210', 'S220', 'S500', 'S500D', 'S505', 'S505D', 'TS15',
 ];
 export const LightBulbs = [
-  'KL110', 'KL120', 'KL125', 'KL130', 'KL135', 'KL50', 'KL60', 'LB100', 'LB110', 'L510', 'L510 Series', 'L530',
-  'L535', 'L630', 'KL400L5', 'KL400L10', 'KL420L5', 'KL430', 'L900', 'L920', 'L930',
+  'KL110', 'KL120', 'KL125', 'KL130', 'KL135', 'KL50', 'KL60', 'LB100', 'LB110', 'L430C', 'L430P',
+  'L510', 'L510 Series', 'L530', 'L535', 'L630', 'KL400L5', 'KL400L10', 'KL420L5', 'KL430', 'L900',
+  'L920', 'L930',
 ];
 export const Unsupported = [
-  'C100', 'C110', 'C210', 'C225', 'C325WB', 'C520WS', 'TC65', 'TC70', 'KH100', 'H100', 'H200', 'KE100',
-  'S200B', 'S200D', 'T100', 'T110', 'T300', 'T310', 'T315',
+  'C100', 'C110', 'C210', 'C220', 'C225', 'C325WB', 'C520WS', 'C720', 'TC65', 'TC70', 'D100C', 'D130',
+  'D230', 'RV20 Max Plus', 'RV30 Max', 'KH100', 'H100', 'H200', 'KE100', 'S200B', 'S200D', 'T100',
+  'T110', 'T300', 'T310', 'T315',
 ];
 
 export interface DescriptorContext {
@@ -134,6 +137,7 @@ export interface CharacteristicDescriptor {
   type: WithUUID<new () => Characteristic>;
   name?: string;
   writable?: boolean;
+  debouncePolls?: number;
   getInitial(context: DescriptorContext): CharacteristicValue;
   getCurrent(context: DescriptorContext): CharacteristicValue;
   applySet?(value: CharacteristicValue, context: DescriptorContext): Promise<void>;